requirements.txt
distorm3
pyelftools